+++ packages/studio-base/src/providers/ExtensionRegistryProvider/useExtensionLoaders.ts
