+++ packages/studio-base/src/providers/ExtensionRegistryProvider/useExtensionPanels.ts
